chore: address deepsource concerns
