+++ src/js/groups/types.ts
@@ -1,4 +1,4 @@
-import { Permissions, UserNested } from "../users/types";
+import { UserNested } from "../users/types";
 
 export type GroupMinimal = { id: string; name?: string };
 
@@ -7,13 +7,18 @@ export type groupNameUpdate = {
     name?: string;
 };
 
-export type groupPermissionsUpdate = {
-    id: string;
-    permissions: Permissions;
-    value: boolean;
-};
-
 export type Group = GroupMinimal & {
     permissions: Permissions;
     users: UserNested[];
+};
+
+export type Permissions = {
+    cancel_job: boolean;
+    create_ref: boolean;
+    create_sample: boolean;
+    modify_hmm: boolean;
+    modify_subtraction: boolean;
+    remove_file: boolean;
+    remove_job: boolean;
+    upload_file: boolean;
 };
